app/pyproject.toml
[tool.poetry]
name = "sample-api-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.118.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.43"}
pydantic-settings = "^2.11.0"
asyncpg = "^0.30.0"
pyplugin-lib = "^0.2.1"
uvicorn = "^0.37.0"


[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
pyright = "^1.1.406"


[tool.poetry.group.development.dependencies]
black = "^25.9.0"
pyright = "^1.1.406"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
